refactor(package): 비즈니스 로직 객체 domain 패키지로 이동
