+++ src/main/kotlin/step3/domain/Car.kt
@@ -1,4 +1,4 @@
-package step3
+package step3.domain
 
 class Car(
     private val engine: ConditionalOperatingCarEngine,
