+++ src/main/kotlin/step3/domain/ConditionalOperatingCarEngine.kt
@@ -1,4 +1,4 @@
-package step3
+package step3.domain
 
 import kotlin.random.Random
