+++ src/main/kotlin/step3/domain/CurrentCarStatus.kt
@@ -1,4 +1,4 @@
-package step3
+package step3.domain
 
 data class CurrentCarStatus(
     val carName: CarName,
