+++ src/main/kotlin/step3/domain/RacingRequirement.kt
@@ -1,4 +1,4 @@
-package step3
+package step3.domain
 
 data class RacingRequirement(
     val carNames: List<CarName>,
