+++ src/main/kotlin/step3/view/ConsoleOutputView.kt
@@ -1,7 +1,7 @@
 package step3.view
 
-import step3.CarName
-import step3.CurrentCarStatus
+import step3.domain.CarName
+import step3.domain.CurrentCarStatus
 
 class ConsoleOutputView : OutputView {
     override fun showResultTitle() {
