+++ src/main/kotlin/step3/view/InputView.kt
@@ -1,6 +1,6 @@
 package step3.view
 
-import step3.RacingRequirement
+import step3.domain.RacingRequirement
 
 interface InputView {
     fun askRequirement(): RacingRequirement
