+++ src/test/kotlin/step3/domain/RacingTest.kt
@@ -1,4 +1,4 @@
-package step3
+package step3.domain
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldMatchInOrder
