+++ src/test/kotlin/step3/test/util.kt
@@ -1,6 +1,6 @@
 package step3.test
 
-import step3.CarName
+import step3.domain.CarName
 
 internal fun String.toCarName(): CarName {
     return CarName(this)
